requirements.txt
pynput==1.7.7
pyautogui==0.9.54
Pillow==11.1.0
pywinauto==0.6.8
comtypes==1.4.8
opencv-python==4.10.0.84
numpy>=1.26.0
pytesseract==0.3.13
anthropic>=1.0.0
pystray==0.19.5
